scaffolding for new structuring subpackage
