+++ beep/structure/arbin.py
@@ -1,0 +1,53 @@
+from datetime import datetime
+
+import pytz
+
+from beep.conversion_schemas import ARBIN_CONFIG
+from beep import logger
+from beep.structure.base import BEEPDatapath
+
+
+class ArbinDatapath(BEEPDatapath):
+
+    @classmethod
+    def from_file(cls, path, metadata_path=None):
+        """
+        """
+        data = pd.read_csv(path)
+        data.rename(str.lower, axis="columns", inplace=True)
+
+        for column, dtype in ARBIN_CONFIG["data_types"].items():
+            if column in data:
+                if not data[column].isnull().values.any():
+                    data[column] = data[column].astype(dtype)
+
+        data.rename(ARBIN_CONFIG["data_columns"], axis="columns", inplace=True)
+
+        metadata_path = metadata_path if metadata_path else path.replace(".csv",
+                                                                         "_Metadata.csv")
+
+        if os.path.exists(metadata_path):
+            metadata = pd.read_csv(metadata_path)
+            metadata.rename(str.lower, axis="columns", inplace=True)
+            metadata.rename(ARBIN_CONFIG["metadata_fields"], axis="columns",
+                            inplace=True)
+            # Note the to_dict, which scrubs numpy typing
+            metadata = {col: item[0] for col, item in
+                        metadata.to_dict("list").items()}
+        else:
+            logger.warning(f"No associated metadata file for Arbin: "
+                           f"'{metadata_path}'. No metadata loaded.")
+            metadata = {}
+
+        # standardizing time format
+        data["date_time_iso"] = data["date_time"].apply(
+            lambda x: datetime.utcfromtimestamp(x).replace(
+                tzinfo=pytz.UTC).isoformat()
+        )
+
+        paths = {
+            "raw": path,
+            "metadata": metadata_path if metadata else None
+        }
+
+        return cls(data, metadata, paths)
